+++ packages/netsuite-adapter/src/suiteapp_config_elements.ts
@@ -16,10 +16,10 @@
 import os from 'os'
 import _ from 'lodash'
 import { BuiltinTypes, Element, ElemID, getChangeData, InstanceElement, isInstanceElement, ModificationChange, ObjectType } from '@salto-io/adapter-api'
-import { NETSUITE, SELECT_OPTION, SETTINGS_PATH, TYPES_PATH } from '../../constants'
-import { ConfigRecord, isSuccessSetConfig, SetConfigRecordsValuesResult, SetConfigType } from './types'
-import { SUITEAPP_CONFIG_TYPES_TO_TYPE_NAMES, DeployResult } from '../../types'
-import { NetsuiteQuery } from '../../query'
+import { NETSUITE, SELECT_OPTION, SETTINGS_PATH, TYPES_PATH } from './constants'
+import { SUITEAPP_CONFIG_TYPES_TO_TYPE_NAMES, DeployResult } from './types'
+import { NetsuiteQuery } from './query'
+import { ConfigRecord, isSuccessSetConfig, SetConfigRecordsValuesResult, SetConfigType } from './client/suiteapp_client/types'
 
 export const getConfigTypes = (): ObjectType[] => ([new ObjectType({
   elemID: new ElemID(NETSUITE, SELECT_OPTION),
